+++ CHANGELOG.md
@@ -1,9 +1,23 @@
+## 1.0.25
+
+* **Fix Notion user text and html getters**
+
+## 1.0.24
+
+* **Handle both cloud and non-cloud jira instances**
+
+## 1.0.23
+
+* **Migrate to new Mixedbread Python SDK**
+* **Support better filtering in jira connector and downloading attachments**
+
+## 1.0.22
+
+* **Fix Notion connector missing database properties fields**
+
 ## 1.0.21
 
 * **Fix Jira connector cloud option not working issue**
-
-## 1.0.20
-
 * **Fix Weaviate connector issue with names being wrongly transformed to match collections naming conventions**
 
 ## 1.0.19